Revert "Squashed commit of the following:"
This reverts commit f4570211a5c304f2c7a50af52904adf8b0ada135.

+++ TombEngine/Game/control/event.h
@@ -8,7 +8,7 @@ namespace TEN::Control::Volumes
 {
 	using Activator = std::variant<
 		std::nullptr_t,
-		int,
+		short,
 		MESH_INFO*,
 		CAMERA_INFO*>;
