+++ TombEngine/Scripting/Include/VarMapVal.h
@@ -9,7 +9,7 @@ struct AI_OBJECT;
 struct ROOM_INFO;
 
 using VarMapVal = std::variant<
-	int,
+	short,
 	std::reference_wrapper<MESH_INFO>,
 	std::reference_wrapper<LevelCameraInfo>,
 	std::reference_wrapper<SinkInfo>,
